Cleanup some pathnames, moving commands into lib
