+++ lib/commands/gamepicker.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder, EmbedBuilder, Collection } = require('discord.js');
-const logger = require('../lib/logger.js');
+const logger = require('../logger.js');
 
 const gamesList = new Collection();
