+++ lib/commands/pfunk.js
