+++ lib/commands/plus_plus.js
@@ -10,7 +10,7 @@ module.exports = {
 			.setTitle('plus-plus Leaderboard')
 			.setDescription('The top twelve users on the plus-plus leaderboard are:');
 
-		const plusplus = require('../data/plusPlus.json');
+		const plusplus = require('../../data/plusPlus.json');
 		console.log(plusplus);
 		const users = Object.keys(plusplus).map(user => { return [ plusplus[user]['id'], plusplus[user]['score'] ] });
 		users.sort((a, b) => { return b[1] - a[1] });
