+++ lib/commands/userlist.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder } = require('discord.js');
-const logger = require('../lib/logger.js');
+const logger = require('../logger.js');
 
 module.exports = {
 	data: new SlashCommandBuilder()
